Meaningless Deserealization removed
I have implemented a class filter before but the vulnerability is not resolved in sonarcloud. I had also thought about implementing it with JSON with the Jackson library but in the original code there is no target class to deserialize, so the last option is to delete the code that deserializes the token object for no reason.